refactor names and remove unused files
