--- src/main/java/fazirul/fyp/dragon/agreement/Agreement.java
@@ -1,4 +1,0 @@
-package fazirul.fyp.dragon.agreement;
-
-public class Agreement {
-}
