--- src/main/java/fazirul/fyp/dragon/agreement/PerNodeAgreement.java
@@ -1,4 +1,0 @@
-package fazirul.fyp.dragon.agreement;
-
-public class PerNodeAgreement {
-}
