+++ src/main/java/fazirul/fyp/dragon/main/Main.java
@@ -1,0 +1,6 @@
+package fazirul.fyp.dragon.main;
+
+public class Main {
+    public static void main(String[] args) {
+    }
+}
