--- src/main/java/fazirul/fyp/dragon/utils/Constants.java
@@ -1,7 +1,0 @@
-package fazirul.fyp.dragon.utils;
-
-public class Constants {
-    public static double VM_DEFAULT_MIPS = 100;
-    public static int HOST_DEFAULT_MIPS = 1000;
-    public static int CLOUDLET_DEFAULT_MI = 1000;
-}
