+++ src/main/java/fazirul/fyp/dragon/utils/Message.java
@@ -1,30 +1,23 @@
 package fazirul.fyp.dragon.utils;
 
-import fazirul.fyp.dragon.election.GlobalData;
+import fazirul.fyp.dragon.dragonDevice.GlobalData;
 import fazirul.fyp.elements.MessageInterface;
-import fazirul.fyp.elements.Node;
-
 import java.time.LocalTime;
-import java.util.HashMap;
-import java.util.HashSet;
 
-// message to be sent
 public class Message implements MessageInterface {
-    private GlobalData data;
-    private int senderID;
-    private LocalTime timestamp;
-    private HashMap<Node, HashSet<Integer>> winnersPerNode;
+    private final GlobalData data;
+    private final int senderID;
+    private final LocalTime timestamp;
 
-    public Message(GlobalData data, int senderID, LocalTime timestamp, HashMap<Node, HashSet<Integer>> winnersPerNode) {
+    public Message(GlobalData data, int senderID, LocalTime timestamp) {
         this.data = data;
         this.senderID = senderID;
         this.timestamp = timestamp;
-        this.winnersPerNode = winnersPerNode;
     }
 
     @Override
     public Message clone() {
-        return new Message(this.data.clone(), this.senderID, this.timestamp, new HashMap<>(winnersPerNode));
+        return new Message(data.clone(), senderID, timestamp);
     }
 
     public int getSenderID() {
@@ -37,9 +30,5 @@ public class Message implements MessageInterface {
 
     public GlobalData getData() {
         return data;
-    }
-
-    public HashSet<Integer> getWinnersForNode(Node n) {
-        return winnersPerNode.get(n);
     }
 }
