--- src/main/java/fazirul/fyp/elements/UtilizationModelAbsolute.java
@@ -1,17 +1,0 @@
-package fazirul.fyp.elements;
-
-import org.cloudbus.cloudsim.utilizationmodels.UtilizationModelAbstract;
-
-public class UtilizationModelAbsolute extends UtilizationModelAbstract {
-    private final double value;
-
-    public UtilizationModelAbsolute(double value) {
-        this.value = value;
-        setUnit(Unit.ABSOLUTE);
-    }
-
-    @Override
-    protected double getUtilizationInternal(double time) {
-        return value;
-    }
-}
